refactor: move all config related code to new config.ts
