+++ module.d.ts
@@ -1,4 +1,4 @@
 declare module "auth:config" {
-  const config: import('./server').AstroAuthConfig;
+  const config: import('./src/config').AstroAuthConfig;
   export default config;
 }
